pyproject.toml
[tool.poetry]
name = "fluxgnn_icml2024"
version = "0.0.8"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.10"

numpy = "<1.24.0"
femio = {version = ">=0.3.3", allow-prereleases = true}
pysiml = {path = "lib/siml/dist/pysiml-0.2.9-py3-none-any.whl"}
PyQt5 = {version = "^5.14.0", optional = true}
mysqlclient = {version = "^1.4.2", optional = true}
SQLAlchemy = {version = "^1.3.4", optional = true}
lxml = "^4.6.3"

[tool.poetry.dev-dependencies]
pytest = "^5.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
